refactor(esbabbler): extract slash command submit into composable
