+++ packages/app/app/components/Message/Model/Message/Input/SlashCommandParameters/Index.vue
@@ -1,22 +1,14 @@
 <script setup lang="ts">
 import type { SlashCommand } from "@/models/message/slashCommands/SlashCommand";
-import type { SlashCommandParameters } from "@/models/message/slashCommands/SlashCommandParameters";
-import type { SlashCommandType } from "@/models/message/slashCommands/SlashCommandType";
 
-import { slashCommandParameterValueSchema } from "@/models/message/slashCommands/SlashCommandParameter";
-import { REQUIRED_ERROR_MESSAGE } from "@/services/message/slashCommands/constants";
 import { SlashCommandDefinitions } from "@/services/message/slashCommands/SlashCommandDefinitionMap";
 import { useSlashCommandStore } from "@/store/message/input/slashCommand";
-import { useRoomStore } from "@/store/message/room";
 
-const roomStore = useRoomStore();
-const { currentRoomId } = storeToRefs(roomStore);
 const slashCommandStore = useSlashCommandStore();
-const { activeParameterNames, activeParameters, focusedIndex, hiddenParameters, lastAddedParameterName, parameterValues, pendingSlashCommand } =
+const { activeParameters, focusedIndex, hiddenParameters, lastAddedParameterName, parameterValues, pendingSlashCommand } =
   storeToRefs(slashCommandStore);
-const { clearPendingSlashCommand, collapseToText, createParameter, deleteParameter: baseDeleteParameter, setErrors, setPendingSlashCommand } =
-  slashCommandStore;
-const executeSlashCommand = useExecuteSlashCommand();
+const { collapseToText, createParameter, deleteParameter: baseDeleteParameter, setPendingSlashCommand } = slashCommandStore;
+const submit = useSubmitSlashCommand();
 const commandTitle = ref(pendingSlashCommand.value?.type ?? "");
 
 watch(pendingSlashCommand, (newPendingSlashCommand) => {
@@ -57,39 +49,6 @@ const updateParameterValue = (name: string, value: string) => {
 const deleteLastParameter = () => {
   const lastIndex = activeParameters.value.length - 1;
   if (lastIndex >= 0) deleteParameter(lastIndex);
-};
-const submit = async () => {
-  if (!pendingSlashCommand.value || !currentRoomId.value) return;
-
-  const missingRequiredParameters = pendingSlashCommand.value.parameters.filter(
-    ({ isRequired, name }) =>
-      isRequired && !slashCommandParameterValueSchema.safeParse(parameterValues.value[name]).success,
-  );
-
-  for (const { isRequired, name } of pendingSlashCommand.value.parameters)
-    if (isRequired)
-      setErrors(
-        name,
-        slashCommandParameterValueSchema.safeParse(parameterValues.value[name]).success ? [] : [REQUIRED_ERROR_MESSAGE],
-      );
-
-  if (missingRequiredParameters.length > 0) {
-    const hiddenMissingParameters = missingRequiredParameters.filter(
-      ({ name }) => !activeParameterNames.value.includes(name),
-    );
-    if (hiddenMissingParameters.length > 0)
-      activeParameterNames.value = [...activeParameterNames.value, ...hiddenMissingParameters.map(({ name }) => name)];
-    return;
-  }
-
-  const payload = {
-    parameterValues: parameterValues.value,
-    type: pendingSlashCommand.value.type,
-  } as {
-    [P in SlashCommandType]: { parameterValues: SlashCommandParameters<P>; type: P };
-  }[SlashCommandType];
-  clearPendingSlashCommand();
-  await executeSlashCommand(payload);
 };
 
 onKeyStroke("Escape", () => collapseToText());
